chore: move benchmarks to a separate project
