+++ .cargo/config.toml
@@ -1,5 +1,5 @@
 [alias]
-benchmark = "bench --features benchmark --package ironfish_rust"
+benchmark = "bench --package benchmarks"
 
 [target.aarch64-unknown-linux-gnu]
 linker = "aarch64-linux-gnu-gcc"
