+++ Cargo.toml
@@ -1,6 +1,7 @@
 [workspace]
 
 members = [
+    "benchmarks",
     "ironfish-mpc",
     "ironfish-phase2",
     "ironfish-rust",
