+++ benchmarks/README.md
@@ -1,0 +1,11 @@
+# benchmarks
+
+A collection of benchmarks and other related performance tests.
+
+These are for development purposes only and do not get shipped with the Node. As such, changes in this package do not require a version bump for releases.
+
+## Usage
+
+`cargo benchmark` - run all benchmarks
+
+`cargo benchmark -- simple` - run benchmarks with "simple" in the name.
